Dominik_Švagera/Závěrečný test/vystupni_test/rf/requirements.txt
robotframework==7.3.2
robotframework-seleniumlibrary==6.8.0
selenium==4.25.0
webdriver-manager==4.0.2
